% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeman.R
\name{aeman}
\alias{aeman}
\title{aeman}
\usage{
aeman(d, line, log10 = TRUE, yaxis, opacity = 1, title = NULL,
  highlight_var, highlight_p, highlighter = "red", color1 = "#AAAAAA",
  color2 = "#4D4D4D", groupcolors, file = "aeman", hgt = 800, wi = 1300)
}
\arguments{
\item{d}{data frame, columns one, two, and three must be Variable, pvalue, and Frame; Group, Shape and Color optional}

\item{line}{optional pvalue threshold to draw red line at}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis, automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0 to 1, useful for dense plots}

\item{title}{optional string for plot title}

\item{highlight_var}{list of variables to highlight}

\item{highlight_p}{pvalue threshold to highlight}

\item{highlighter}{color to highlight}

\item{color1}{first alternating color}

\item{color2}{second alternating color}

\item{groupcolors}{named list of colors for data in 'Color' column}

\item{file}{file name of saved image}

\item{hgt}{height of plot in pixels}

\item{wi}{width of plot in pixels}

\item{ext}{file type to save, "gif" or "html"}
}
\value{
.gif or .html file
}
\description{
Create animated Manhattan plots for EWAS
Note: There is an issue with dev.off() if using RStudio
Dependencies: ggplot2
Suggested: RColorBrewer
}
\examples{
aeman(d, groups, line, title=NULL, file="eman", hgt=1300, wi=800, )
}
\seealso{
\code{\link{eman}}, \code{\link{ieman}}, \code{\link{agman}}, \code{\link{apheman}}

Other EWAS functions: \code{\link{eman}},
  \code{\link{ieman}}

Other animated plotting functions: \code{\link{agman}},
  \code{\link{apheman}}
}
