% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agman.R
\name{agman}
\alias{agman}
\title{agman}
\usage{
agman(d, line, log10 = TRUE, yaxis, opacity = 1, annotate_snp, annotate_p,
  highlight_snp, highlight_p, highlighter = "red", title = NULL,
  chrcolor1 = "#AAAAAA", chrcolor2 = "#4D4D4D", groupcolors,
  background = "variegated", chrblocks = FALSE, file = "agman",
  ext = "gif", hgt = 800, wi = 1300)
}
\arguments{
\item{d}{data frame, if not plato or plink format, must contain SNP, CHR, POS, pvalue, Frame columns, optional Shape and Color}

\item{line}{optional pvalue threshold to draw red line at}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis, automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0 to 1, useful for dense plots}

\item{annotate_snp}{list of SNPs to annotate}

\item{annotate_p}{pvalue threshold to annotate}

\item{highlight_snp}{list of SNPs to highlight}

\item{highlight_p}{pvalue threshold to highlight}

\item{highlighter}{color to highlight}

\item{title}{optional string for plot title}

\item{chrcolor1}{first alternating color for chromosome}

\item{chrcolor2}{second alternating color for chromosome}

\item{groupcolors}{named list of colors for data in 'Color' column}

\item{background}{variegated or white}

\item{chrblocks}{boolean, turns on x-axis group marker blocks}

\item{file}{file name of saved image}

\item{ext}{file type to save, "gif" or "mp4"}

\item{hgt}{height of plot in pixels}

\item{wi}{width of plot in pixels}
}
\value{
gif or mp4 file
}
\description{
Create Manhattan plots for GWAS
}
\examples{
#To use the animated plot function, we need to add an animation 'Frame' column to our data
#In this case we will imagine that we've run a GWAS using additive, dominant, and recessive models
#and want to highlight a SNP of interest to see how the p-value changes
data(gwas)
agman(d=gwas, line=0.0005, highlight_snp="rs4204", chrcolor1="#D4CAA0", chrcolor2="#B3BC92", highlighter="black",opacity=0.7, wi=750, hgt=500)
}
\seealso{
\code{\link{gman}}, \code{\link{igman}}, \code{\link{apheman}}, \code{\link{aeman}}

Other GWAS functions: \code{\link{gman}},
  \code{\link{igman}}

Other animated plotting functions: \code{\link{aeman}},
  \code{\link{apheman}}
}
