% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqqunif.R
\name{aqqunif}
\alias{aqqunif}
\title{aqqunif}
\usage{
aqqunif(d, CI = 0.95, opacity = 1, groupcolors, splitby = NULL,
  highlight_p, highlight_name, annotate_p, annotate_name, highlighter = "red",
  line, background, title, file = "aqqunif", ext = "gif", wi = 800,
  hgt = 800)
}
\arguments{
\item{d}{dataframe with at least two columns, p-value and Frame; Color, Shape, and Name optional}

\item{CI}{two-sided confidence interval, default 0.95}

\item{opacity}{point opacity, default 1}

\item{groupcolors}{named vector of colors corresponding to data in Group column}

\item{splitby}{if data contains Color and/or Shape, indicate variable(s) by which the data should be subsetted for calculating CIs}

\item{highlight_p}{p-value threshold to highlight}

\item{highlight_name}{vector of names to highlight, dataframe must include a Name column}

\item{annotate_p}{p-value threshold to annotate, dataframe must include a Name column}

\item{annotate_name}{vector of names to annotate, dataframe must include a Name column}

\item{highlighter}{highlighter color}

\item{line}{draw a red line at pvalue threshold (observed)}

\item{background}{can change to "white"}

\item{title}{plot title}

\item{file}{filename}

\item{ext}{file type to save, "gif" or "mp4"}

\item{wi}{width of plot in pixels}

\item{hgt}{height of plot in pixels}
}
\value{
png image(s)
}
\description{
Create animated qqplots with an assumed uniform distribution
}
\examples{
data(gwas)
aqqunif(d=gwas[, 4:5], opacity=0.6, splitby="Color")
}
