% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bus-path.R
\name{bus_path}
\alias{bus_path}
\title{Bus path details}
\format{
A list with 4 elements:
\describe{
\item{route}{The route code.}
\item{name}{The route name.}
\item{shape}{The coordinate path of the line in both directions.}
\item{stops}{The coordinates of stops on a given line.}
}
}
\usage{
bus_path(route = NULL, date = NULL)
}
\arguments{
\item{route}{Bus route variant, e.g.: 70, 10A, 10Av1.}

\item{date}{Date for which to retrieve route and stop information.}
}
\description{
For a given date, returns the set of ordered latitude/longitude points along
a route variant along with the list of stops served. Useful for creating
maps.
}
