% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bus-position.R
\name{bus_position}
\alias{bus_position}
\title{Bus position}
\format{
A tibble with 13 variables:
\describe{
\item{vehicle}{Unique identifier. Usually visible on the bus itself.}
\item{lat}{Last reported Latitude of the bus.}
\item{lon}{Last reported Longitude of the bus.}
\item{dist}{Distance in meters from provided coordinates.}
\item{deviation}{Deviation, in minutes, from schedule.}
\item{time}{Date and time (EST) of last position update.}
\item{trip}{Unique trip ID. Used with the schedule-related functions.}
\item{route}{Base route name as shown on the bus.}
\item{direction}{General direction of the trip, not the bus itself }
\item{head}{Destination of the bus.}
\item{start}{Scheduled start time (EST) of the bus's current trip.}
\item{end}{Scheduled end time (EST) of the bus's current trip.}
\item{block}{}
}
}
\source{
\url{https://api.wmata.com/Bus.svc/json/jBusPositions}
}
\usage{
bus_position(route = NULL, lat = NULL, lon = NULL, radius = 1000)
}
\arguments{
\item{route}{Base bus route, e.g.: 70, 10A.}

\item{lat}{Center point Latitude.}

\item{lon}{Center point Longitude.}

\item{radius}{Meters to include in the search area for \code{lat} and \code{lon}.}
}
\description{
Returns bus positions for the given route, with an optional search radius. If
no parameters are specified, all bus positions are returned.
}
\details{
Note that the \code{route} parameter accepts only base route names and no
variations, i.e.: use 10A instead of 10Av1 or 10Av2.
}
