% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bus-routes.R
\name{bus_routes}
\alias{bus_routes}
\title{Bus routes}
\format{
A tibble with 3 variables:
\describe{
\item{route}{Unique identifier for a given route variant.}
\item{name}{Descriptive name of the route variant.}
\item{description}{Denotes the route variant’s grouping – lines are a
combination of routes which lie in the same corridor and which have
significant portions of their paths along the same roadways.}
}
}
\source{
\url{https://api.wmata.com/Bus.svc/json/jRoutes}
}
\usage{
bus_routes()
}
\description{
Returns a list of all bus route variants (patterns). For example, the 10A and
10Av1 are the same route, but may stop at slightly different locations.
}
