% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bus-schedule.R
\name{bus_schedule}
\alias{bus_schedule}
\title{Bus route and stop methods}
\format{
A tibble with 10 variables:
\describe{
\item{route_id}{The bus route ID}
\item{direction}{General route direction}
\item{trip_head}{Destination of the bus.}
\item{trip_id}{Trip identifier.}
\item{start_time}{Scheduled start date and time (EST) for this trip.}
\item{end_time}{Scheduled end date and time (EST) for this trip.}
\item{stop_seq}{Order of the stop in the sequence of StopTimes.}
\item{stop_id}{7-digit regional ID. If unavailable, 0 or NULL.}
\item{stop_name}{Stop name. May be different from what is displayed.}
\item{stop_time}{Scheduled departure date and time (EST) from this stop.}
}
}
\source{
\url{https://api.wmata.com/Bus.svc/json/jRouteSchedule}
}
\usage{
bus_schedule(route = NULL, variants = TRUE, date = NULL)
}
\arguments{
\item{route}{Bus route variant, e.g.: 70, 10A, 10Av1, etc.}

\item{variants}{Should variations of \code{route} be included?. For example, if
B30 is specified and \code{TRUE} (default), data for all variations of B30 such
as B30v1, B30v2, etc. will be returned.}

\item{date}{Date for which to retrieve route and stop information.}
}
\description{
Returns schedules for a given route variant for a given date.
}
