% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bus-stops.R
\name{bus_stops}
\alias{bus_stops}
\title{Rail stop search}
\format{
A tibble with 6 variables:
\describe{
\item{stop}{Bus stop ID}
\item{name}{Bus stop name}
\item{lon}{Latitude.}
\item{lat}{Longitude.}
\item{distance}{Distance in meters from provided coordinates.}
\item{routes}{Character vector of routes services by stop.}
}
}
\usage{
bus_stops(lat = NULL, lon = NULL, radius = NULL)
}
\arguments{
\item{lat}{Center point Latitude.}

\item{lon}{Center point Longitude}

\item{radius}{Radius (meters) to include in the search area. If \code{NULL}, while
\code{lat} and \code{lon} are supplied, all stations are returned.}
}
\description{
Returns a list of nearby bus stops based on latitude, longitude, and radius.
Omit all parameters to retrieve a list of all stops.
}
