% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_p.R
\name{generate_p}
\alias{generate_p}
\title{generate_p()}
\usage{
generate_p(color = FALSE, shape = FALSE)
}
\arguments{
\item{shape}{Boolean include Shape column}

\item{group}{Boolean include Group column}
}
\value{
png image(s)
}
\description{
Generate dummy data
}
\examples{
generate_p(color, shape)
}
