% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ieman.R
\name{ieman}
\alias{ieman}
\title{ieman}
\usage{
ieman(d, line, log10 = TRUE, yaxis, opacity = 1, title = NULL,
  highlight_var, highlight_p, highlighter = "red", color1 = "#AAAAAA",
  color2 = "#4D4D4D", groupcolors, db, moreinfo = FALSE, file = "ieman",
  hgt = 7, wi = 12, bigrender = FALSE)
}
\arguments{
\item{d}{data frame, columns one and two must be Variable and pvalue; Group, Shape, Color, and Info optional}

\item{line}{optional pvalue threshold to draw red line at}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis, automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0 to 1, useful for dense plots}

\item{title}{optional string for plot title}

\item{highlight_var}{list of variables to highlight}

\item{highlight_p}{pvalue threshold to highlight}

\item{highlighter}{color to highlight}

\item{color1}{first alternating color}

\item{color2}{second alternating color}

\item{groupcolors}{named list of colors for data in 'Color' column}

\item{db}{query address, ex. "https://www.google.com/search?q"}

\item{moreinfo}{includes more information on hover, refers to Info column}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{bigrender}{can set to TRUE for big plots (~50000 rows) that produce huge input lookup error}
}
\value{
png image(s)
}
\description{
Create interactive Manhattan plots for EWAS
Note: There is an issue with dev.off() if using RStudio
Dependencies: ggplot2
Suggested: RColorBrewer
}
\examples{
ieman(d, groups, line, title=NULL, morecolors=FALSE, file="eman", hgt=7, wi=12, res=300 )
}
\seealso{
\code{\link{eman}}, \code{\link{aeman}}, \code{\link{igman}}, \code{\link{ipheman}}

Other EWAS functions: \code{\link{aeman}},
  \code{\link{eman}}
}
