% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igman.R
\name{igman}
\alias{igman}
\title{igman}
\usage{
igman(d, format = "plotman", line, log10 = TRUE, yaxis, opacity = 1,
  highlight_snp, highlight_p, highlighter = "red", title = NULL,
  chrcolor1 = "#AAAAAA", chrcolor2 = "#4D4D4D", groupcolors, db,
  moreinfo = FALSE, file = "igman", hgt = 7, wi = 12)
}
\arguments{
\item{d}{data frame, if not plato or plink format, must contain SNP, CHR, POS, pvalue columns, optional Shape, Color, and Info}

\item{format}{format of input}

\item{line}{optional pvalue threshold to draw red line at}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis, automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0 to 1, useful for dense plots}

\item{highlight_snp}{list of snps to highlight}

\item{highlight_p}{pvalue threshold to highlight}

\item{highlighter}{color to highlight}

\item{title}{optional string for plot title}

\item{chrcolor1}{first alternating color for chromosome}

\item{chrcolor2}{second alternating color for chromosome}

\item{groupcolors}{named list of colors for data in 'Color' column}

\item{db}{choose database to connect to (dbSNP or GWAScatalog)}

\item{moreinfo}{includes more information on hover, refers to Info column}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}
}
\value{
html file
}
\description{
Create Interactive Manhattan plots for GWAS
Dependencies: ggplot2, ggiraph
Suggested: RColorBrewer
}
\examples{
igman(d, format, line, log10, yaxis, title, chrcolor1, chrcolor2, groupcolors, db, moreinfo, file, hgt, wi)
}
