% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METRO2SumStat.R
\name{METRO2SumStat}
\alias{METRO2SumStat}
\title{METRO with summary level expression data and summary level GWAS data}
\usage{
METRO2SumStat(
  eQTLzscores,
  eQTLLDs,
  GWASzscores,
  GWASLD,
  ns,
  n,
  nu = 0.8,
  hthre = 0.002,
  maxIter = 1000,
  tol = 0.001,
  verbose = FALSE
)
}
\arguments{
\item{eQTLzscores}{A pxM numeric matrix of marginal zscores for p SNPs
from M eQTL studies conducted in different ancestries.}

\item{eQTLLDs}{M-list of SNP-SNP correlation (LD) matrices estimated from
reference panels with individuals in the same ancestries as expression
studies. The order has to match across eQTL zscores, eQTL LD matrices, 
and the number of individuals.}

\item{GWASzscores}{A numeric vector of marginal zscores for p SNPs from GWAS.}

\item{GWASLD}{A SNP-SNP correlation (LD) matrix estimated from a reference
panel with individuals in the same ancestry as GWAS data.}

\item{ns}{A numeric vector. The number of individuals in each expression study.
The order has to match across eQTL zscores, eQTL LD matrices, and the
number of individuals.}

\item{n}{The number of individuals in GWAS data.}

\item{nu}{The shrinkage intensity parameter ranging from 0 to 1 for the LD
matrix. A smaller value indicates higher intensity.}

\item{hthre}{A numeric scalar. Heritability threshold to determine whether
cis-SNPs are predictive of gene expression.}

\item{maxIter}{An integer. Maximum number of iterations for PX-EM algorithm.}

\item{tol}{A numeric scalar. Convergence tolerance for PX-EM algorithm.}

\item{verbose}{A logical scalar. Whether to show verbose information.}
}
\value{
\code{METRO2SumStat} returns a list of estimated parameters and 
   statistics from METRO:
\item{alpha}{A numeric scalar of expression-on-outcome effect.}
\item{weights}{A numeric vector of weights from multiple genetic ancestries.}
\item{beta}{A numeric vector of genetic effects on gene expression among GWAS
   individuals.}
\item{LRTStat}{Likelihood ratio test statistic.}
\item{df}{Degree of freedoms of the reference chi2 distribution for LRTStat.}
\item{pvalueLRT}{P-value of the likelihood ratio test.}
\item{sigma2beta}{A numeric vector of variance estimates of cis-SNP effects
   in different genetic ancestries.}
\item{sigma2m}{A numeric vector of residual variance estimates in gene 
   expression models.}
\item{sigma2}{The residual variance estiamte in the GWAS-expression 
   association model.}
\item{h2m}{A numeric vector of cis-SNP heritabilities of gene expression in
   different genetic ancestries.}
\item{h2y}{The proportion of variance in GWAS outcome explained by gene
   expression.}
\item{p}{The number of cis-SNPs.}
\item{M}{The number of genetic ancesties.}
\item{nullLoglik}{The log-likelihood under the null.}
\item{altLoglik}{The log-likelihood under the alternative.}
\item{elapsedTime}{Time (seconds) used to run METRO.}
}
\description{
METRO with summary level expression data and summary level GWAS data
}
