% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metro_lines}
\alias{metro_lines}
\title{All WMATA rail lines}
\format{
A data frame with 95 rows and 10 variables:
\describe{
\item{line}{Two-letter abbreviation for the line.}
\item{name}{Full name of line color.}
\item{start}{Start station code.}
\item{end}{End station code.}
...
}
}
\source{
\url{https://api.wmata.com/Rail.svc/json/jLines}
}
\usage{
metro_lines
}
\description{
All WMATA rail lines
}
\keyword{datasets}
