% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metro_stations}
\alias{metro_stations}
\title{All WMATA rail stations}
\format{
A data frame with 95 rows and 10 variables:
\describe{
\item{station}{Station code.}
\item{name}{Station name.}
\item{txfer}{For stations with multiple platforms, the additional code.}
\item{lines}{Vector of line codes services by station.}
\item{lat}{Latitude.}
\item{lon}{Longitude.}
\item{street}{Street address (for GPS use).}
\item{city}{City.}
\item{state}{State (abbreviated).}
\item{zip}{Zip code.}
...
}
}
\usage{
metro_stations
}
\description{
All WMATA rail stations
}
\keyword{datasets}
