% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next-train.R
\name{next_train}
\alias{next_train}
\title{Next train}
\format{
A tibble with 9 variables:
\describe{
\item{order}{The order in which trains will arrive at the station.}
\item{station}{Station code for where the train is arriving.}
\item{name}{Full name of the station where the train is arriving.}
\item{min}{Minutes until arrival: "ARR" converted to 0 and "BRD" to -1.}
\item{to_station}{Abbreviated \emph{final} destination for a train.}
\item{to_name}{Destination station code.}
\item{group}{Denotes the track this train is on but not Track "1" or "2".}
\item{line}{Two-letter abbreviation for the line}
\item{cars}{Number of cars on a train, usually 6 or 8.}
}
}
\source{
\url{https://api.wmata.com/StationPrediction.svc/json/GetPrediction/}
}
\usage{
next_train(station = NULL)
}
\arguments{
\item{station}{Character vector of station names or \code{NULL} for all.}
}
\description{
Returns next train arrival information for one or more stations. Will return
an empty set of results when no predictions are available. Use \code{NULL} for the
\code{station} parameter to return predictions for all stations.
}
\details{
For terminal stations (e.g.: Greenbelt, Shady Grove, etc.), predictions may
be displayed twice.

Some stations have two platforms (e.g.: Gallery Place, Fort Totten, L'Enfant
Plaza, and Metro Center). To retrieve complete predictions for these
stations, be sure to pass in both StationCodes.

For trains with no passengers, the DestinationName will be "No Passenger".
}
