% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{no_data_now}
\alias{no_data_now}
\title{Check if data has length and rows}
\usage{
no_data_now(x)
}
\arguments{
\item{x}{A data frame or list to check}
}
\description{
Check if data has length and rows
}
\examples{
no_data_now(data.frame())
no_data_now(data.frame(a = 1))
}
