% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-parking.R
\name{parking_spots}
\alias{parking_spots}
\alias{parking_cost}
\title{Parking information at Metro stations}
\usage{
parking_spots(line = NULL)

parking_cost(line = NULL)
}
\arguments{
\item{line}{Two-letter line code abbreviation, see \link{lines} or \code{\link[=rail_lines]{rail_lines()}}.}
}
\description{
Parking information at Metro stations
}
