% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheman.R
\name{pheman}
\alias{pheman}
\title{pheman}
\usage{
pheman(d, phegroup, line, log10 = TRUE, yaxis, opacity = 1, annotate_snp,
  annotate_p, highlight_snp, highlight_p, highlighter = "red", title = NULL,
  chrcolor1 = "#AAAAAA", chrcolor2 = "#4D4D4D", groupcolors,
  file = "pheman", hgt = 7, wi = 12, res = 300)
}
\arguments{
\item{d}{data frame, if not plato or plink format, must contain PHE, SNP, CHR, POS, pvalue, columns, optional Shape}

\item{phegroup}{optional grouping file for phenotypes, must contain PHE and Group columns}

\item{line}{optional pvalue threshold to draw red line at}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis, automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0-1, useful for dense plots}

\item{annotate_snp}{list of SNPs to annotate}

\item{annotate_p}{pvalue threshold to annotate}

\item{highlight_snp}{list of snps to highlight}

\item{highlight_p}{pvalue threshold to highlight}

\item{highlighter}{color to highlight}

\item{title}{optional string for plot title}

\item{chrcolor1}{first alternating color for chromosome}

\item{chrcolor2}{second alternating color for chromosome}

\item{groupcolors}{named list of colors for data in 'PHE' or 'Group' column}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{res}{resolution of plot in pixels per inch}
}
\value{
png image
}
\description{
Create Manhattan plots for PheWAS
Note: There is an issue with dev.off() if using RStudio
Dependencies: ggplot2
Suggested: RColorBrewer, ggrepel
}
\examples{
#Generate some data
#In this case we'd like to also see the p-value when we hover over a point, so we'll add an 'Info' column to the data
#We'd also like to search dbSNP when we click on a point
dat <- data.frame(PHE=rep(paste("Pheno", seq(1:5), sep="") , length.out=5000, each=1),
                 SNP=paste("rs", seq(1:5000), sep=""),
                 CHR=rep(c(1:22, "X", "Y"), length.out=5000, each=200),
                 POS=rep(seq(1, 10000, by = 200), length.out=5000),
                 pvalue=runif(n=5000))
dat$Info <- paste0("p-value:", signif(dat$pvalue, digits=3))
#Plot
ipheman(d=dat, moreinfo = TRUE, db="dbSNP", line=0.001)
}
\seealso{
\code{\link{ipheman}}, \code{\link{apheman}}, \code{\link{gman}}, \code{\link{eman}}

Other PheWAS functions: \code{\link{apheman}},
  \code{\link{ipheman}}

Other static plotting functions: \code{\link{eman}},
  \code{\link{gman}}
}
