% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas-data.R
\docType{data}
\name{phewas}
\alias{phewas}
\title{Example PheWAS data}
\format{An object of class data.frame}
\usage{
data(phewas)
}
\description{
Example PheWAS data
}
\examples{
data(phewas)
}
\keyword{datasets}
