% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrance.R
\name{rail_entrance}
\alias{rail_entrance}
\title{Rail station entrances}
\usage{
rail_entrance(lat = NULL, lon = NULL, radius = NULL)
}
\arguments{
\item{lat}{Center point Latitude.}

\item{lon}{Center point Longitude}

\item{radius}{Radius (meters) to include in the search area. If \code{NULL}, while
\code{lat} and \code{lon} are supplied, all stations are returned.}
}
\description{
Returns a list of nearby station entrances based on latitude, longitude, and
radius (meters). Omit search parameters to return all station entrances.
}
