% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rail-path.R
\name{rail_path}
\alias{rail_path}
\alias{path_distance}
\title{Path between rail stations}
\usage{
rail_path(from, to)

path_distance(from, to)
}
\arguments{
\item{from}{Station code for the origin station.}

\item{to}{Station code for the destination station.}
}
\description{
Returns a set of ordered stations and distances between two stations on the
same line.
}
\details{
Note that this method is not suitable on its own as a pathfinding solution
between stations. Distance is converted from feet to round meters.
}
