% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\name{rail_times}
\alias{rail_times}
\title{Rail station times}
\usage{
rail_times(station = NULL, dates = TRUE, numeric = FALSE)
}
\arguments{
\item{station}{Station code, see \link{stations} or \code{\link[=rail_stations]{rail_stations()}}.}

\item{dates}{Should daily hours be converted to dates for current week?}

\item{numeric}{Should times be converted to doubles? (e.g., 09:30 to 9.5)}
}
\description{
Returns opening and scheduled first/last train times based on a given
station code. Omit the station to return timing information for all
stations.
}
\details{
Note that for stations with multiple platforms (e.g.: Metro Center, L'Enfant
Plaza, etc.), a distinct call is required for each \code{station} to retrieve
the full set of train times at such stations.
}
