% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-routes.R
\name{standard_routes}
\alias{standard_routes}
\title{Standard routes}
\usage{
standard_routes()
}
\description{
Returns an ordered list of mostly revenue (and some lead) track circuits,
arranged by line and track number. This data does not change frequently and
should be cached for a reasonable amount of time.
}
\details{
Please refer to \href{https://developer.wmata.com/TrainPositionsFAQ}{this page}
for additional details.
}
