% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-circuits.R
\name{track_circuits}
\alias{track_circuits}
\title{Track circuits}
\usage{
track_circuits()
}
\description{
Returns a list of all track circuits including those on pocket tracks and
crossovers. Each track circuit may include references to its right and left
neighbors.
}
\details{
Please refer to \href{https://developer.wmata.com/TrainPositionsFAQ}{this page}
for additional details.
}
