% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-position.R
\name{train_positions}
\alias{train_positions}
\title{Live train position}
\usage{
train_positions()
}
\description{
Returns uniquely identifiable trains in service and what track circuits they
currently occupy. Will return an empty set of results when no positions are
available. Data is refreshed once every 7-10 seconds.
}
\details{
Please refer to \href{https://developer.wmata.com/TrainPositionsFAQ}{this page}
for additional details.
}
