% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wmata_api}
\alias{wmata_api}
\title{Call the WMATA API}
\usage{
wmata_api(path, query = NULL)
}
\arguments{
\item{path}{The API endpoint path, passed to \code{\link[httr:modify_url]{httr::modify_url()}}.}

\item{query}{Additional queries also passed, possibly your key if need be.}
}
\value{
A single JSON string.
}
\description{
Use \code{\link[httr:GET]{httr::GET()}} to make a request to the WMATA API and return the unparsed
JSON string. This function is used in others devoted to a
specific APIs and endpoints.
}
\details{
This function was modified from zamorarr's wmata on GitHub:
\url{https://github.com/zamorarr/wmata/blob/master/R/api.r}
}
\examples{
\dontrun{
wmata_api("/Rail.svc/json/jLines", query = list(LineCode = "RD"))
}
}
\keyword{internal}
