% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{wmata_key}
\alias{wmata_key}
\alias{browse_wmata_key}
\title{Get your WMATA API key}
\usage{
wmata_key()

browse_wmata_key()
}
\description{
All calls to the WMATA API must be accompanied by a personal API key. A free
key can be obtained by subscribing to the default tier. See
\code{\link[=browse_wmata_key]{browse_wmata_key()}} for more information on getting a key and storing on
your system.
}
\details{
From the WMATA developer website:
Default tier sufficient for most casual developers. Rate limited to 10
calls/second and 50,000 calls per day. This product contains 8 APIs:
\itemize{
\item Bus Route and Stop Methods
\item GTFS
\item Incidents
\item Misc Methods
\item Rail Station Information
\item Real-Time Bus Predictions
\item Real-Time Rail Predictions
\item Train Positions
}
}
