% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeman.R
\name{aeman}
\alias{aeman}
\title{aeman}
\usage{
aeman(d, line, log10 = TRUE, yaxis, opacity = 1, title = NULL,
  annotate_var, annotate_p, highlight_var, highlight_p, highlighter = "red",
  color1 = "#AAAAAA", color2 = "#4D4D4D", groupcolors,
  background = "variegated", grpblocks = FALSE, file = "aeman",
  ext = "gif", hgt = 800, wi = 1300)
}
\arguments{
\item{d}{data frame, columns one, two, and three must be Variable, pvalue, Group, and Frame; Shape and Color optional}

\item{line}{optional pvalue threshold to draw red line at}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis, automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0 to 1, useful for dense plots}

\item{title}{optional string for plot title}

\item{annotate_var}{vector of variables to annotate}

\item{annotate_p}{pvalue threshold to annotate}

\item{highlight_var}{vector of variables to highlight}

\item{highlight_p}{pvalue threshold to highlight}

\item{highlighter}{color to highlight}

\item{color1}{first alternating color}

\item{color2}{second alternating color}

\item{groupcolors}{named vector of colors for data in 'Color' column}

\item{background}{variegated or white}

\item{grpblocks}{boolean, turns on x-axis group marker blocks}

\item{file}{file name of saved image}

\item{ext}{file type to save, "gif" or "mp4"}

\item{hgt}{height of plot in pixels}

\item{wi}{width of plot in pixels}
}
\value{
.gif or .mp4 file
}
\description{
Create animated Manhattan plots for EWAS
}
\examples{
data(ewas)
ewas$Frame <- ewas$Group
aeman(ewas, line=0.001, highlight_p=0.001, annotate_p=0.001, highlighter="green", title="EWAS Example:")
}
\seealso{
\code{\link{eman}}, \code{\link{ieman}}, \code{\link{agman}}, \code{\link{apheman}}

Other EWAS functions: \code{\link{eman}},
  \code{\link{ieman}}

Other animated plotting functions: \code{\link{agman}},
  \code{\link{apheman}}
}
