% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agman.R
\name{agman}
\alias{agman}
\title{agman}
\usage{
agman(d, format = "plotman", line, log10 = TRUE, yaxis, opacity = 1,
  highlight_snp, highlight_p, highlighter = "red", title = NULL,
  chrcolor1 = "#AAAAAA", chrcolor2 = "#4D4D4D", groupcolors,
  file = "agman", ext = "gif", hgt = 800, wi = 1300)
}
\arguments{
\item{d}{data frame, if not plato or plink format, must contain SNP, CHR, POS, pvalue, Frame columns, optional Shape and Color}

\item{format}{format of input}

\item{line}{optional pvalue threshold to draw red line at}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis, automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0 to 1, useful for dense plots}

\item{highlight_snp}{list of snps to highlight}

\item{highlight_p}{pvalue threshold to highlight}

\item{highlighter}{color to highlight}

\item{title}{optional string for plot title}

\item{chrcolor1}{first alternating color for chromosome}

\item{chrcolor2}{second alternating color for chromosome}

\item{groupcolors}{named list of colors for data in 'Color' column}

\item{file}{file name of saved image}

\item{ext}{file type to save, "gif" or "html"}

\item{hgt}{height of plot in pixels}

\item{wi}{width of plot in pixels}
}
\value{
png image
}
\description{
Create Manhattan plots for GWAS
Note: There is an issue with dev.off() if using RStudio
Dependencies: ggplot2, gganimate
Suggested: RColorBrewer
}
\examples{
agman(d, format, line, log10, yaxis, opacity, title, chrcolor1, chrcolor2, groupcolors, file, ext, hgt, wi)
}
