% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bus-departs.R
\name{bus_departs}
\alias{bus_departs}
\title{Bus Route and Stop Methods}
\format{
A tibble with 1 row per bus departure and 13 variables:
\describe{
\item{RouteID}{Bus route variant identifier (pattern). This variant can be
used in several other bus methods which accept variants. Note that
customers will never see anything other than the base route name, so
variants 10A, 10Av1, 10Av2, etc. will be displayed as 10A on the bus.}
\item{ScheduleTime}{Date and time (UTC) when the bus is scheduled to stop
at this location.}
\item{TripID}{Trip identifier. This can be correlated with the data in our
bus schedule information as well as bus positions.}
\item{TripDirectionText}{General direction of the trip (e.g.: NORTH, SOUTH,
EAST, WEST).}
\item{TripHeadsign}{Destination of the bus.}
\item{StartTime}{Scheduled start date and time (UTC) for this trip.}
\item{EndTime}{Scheduled end date and time (UTC) for this trip.}
}
}
\usage{
bus_departs(StopID, Date = NULL)
}
\arguments{
\item{StopID}{7-digit regional stop ID.}

\item{Date}{(Optional) Date for which to retrieve route and stop information.}
}
\value{
Data frame containing scheduled arrival information
}
\description{
Returns a set of buses scheduled at a stop for a given date.
}
\examples{
\dontrun{
bus_departs(2000474, "2021-02-16")
}
}
\seealso{
\url{https://developer.wmata.com/docs/services/54763629281d83086473f231/operations/5476362a281d830c946a3d6c/console}

Other Bus Route and Stop Methods: 
\code{\link{bus_path}()},
\code{\link{bus_position}()},
\code{\link{bus_routes}()},
\code{\link{bus_schedule}()},
\code{\link{bus_stops}()}
}
\concept{Bus Route and Stop Methods}
