% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bus-departs.R
\name{bus_departs}
\alias{bus_departs}
\title{Bus route and stop methods}
\format{
A tibble with 13 variables:
\describe{
\item{route}{Bus route variant identifier (pattern).}
\item{depart_time}{Date and time (EST) when the bus is scheduled to stop at
this location.}
\item{trip_id}{Trip identifier.}
\item{direction}{General direction of the trip.}
\item{trip_head}{Destination of the bus.}
\item{start_time}{Scheduled start date and time (EST) for this trip.}
\item{end_time}{Scheduled end date and time (EST) for this trip.}
}
}
\usage{
bus_departs(stop = NULL, date = NULL)
}
\arguments{
\item{stop}{7-digit regional stop ID.}

\item{date}{Date for which to retrieve route and stop information.}
}
\description{
Returns schedules for a given route variant for a given date.
}
