% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bus-incidents.R
\name{bus_incidents}
\alias{bus_incidents}
\title{Bus incidents}
\usage{
bus_incidents(route = NULL)
}
\arguments{
\item{route}{Base bus route; variations are not recognized. Omit the Route to
return all reported items (default).}
}
\description{
Returns a set of reported bus incidents/delays for a given Route.
}
\details{
Note that the Route parameter accepts only base route names and no
variations, i.e.: use 10A instead of 10Av1 and 10Av2.
}
