% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevator-incidents.R
\name{elevator_incidents}
\alias{elevator_incidents}
\title{Bus incidents}
\source{
\url{https://api.wmata.com/Incidents.svc/json/ElevatorIncidents}
}
\usage{
elevator_incidents(station = NULL)
}
\arguments{
\item{station}{Station code. Omit to return all reported outages (default).}
}
\description{
Returns a list of reported elevator and escalator outages at a given station.
}
\details{
Note that for stations with multiple platforms and therefore \code{station} codes
(e.g.: Metro Center, L'Enfant Plaza, etc.), a distinct call is required for
each code.
}
