% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eman.R, R/test.R
\name{eman}
\alias{eman}
\alias{eman}
\title{eman}
\usage{
eman(d, line, log10 = TRUE, yaxis, opacity = 1, title = NULL,
  annotate_var, annotate_p, highlight_var, highlight_p, highlighter = "red",
  color1 = "#AAAAAA", color2 = "#4D4D4D", groupcolors, file = "eman",
  hgt = 7, wi = 12, res = 300)

eman(d, line, log10 = TRUE, yaxis, opacity = 1, title = NULL,
  annotate_var, annotate_p, highlight_var, highlight_p, highlighter = "red",
  color1 = "#AAAAAA", color2 = "#4D4D4D", groupcolors, file = "eman",
  hgt = 7, wi = 12, res = 300)
}
\arguments{
\item{d}{data frame, columns one and two must be Variable and pvalue, Group, Shape and Color optional}

\item{line}{optional pvalue threshold to draw red line at}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis, automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0 to 1, useful for dense plots}

\item{title}{optional string for plot title}

\item{annotate_var}{list of variables to annotate}

\item{annotate_p}{pvalue threshold to annotate}

\item{highlight_var}{list of variables to highlight}

\item{highlight_p}{pvalue threshold to highlight}

\item{highlighter}{color to highlight}

\item{color1}{first alternating color}

\item{color2}{second alternating color}

\item{groupcolors}{named list of colors for data in 'Color' column}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{res}{resolution of plot in pixels per inch}

\item{d}{data frame, columns one and two must be Variable and pvalue, Group, Shape and Color optional}

\item{line}{optional pvalue threshold to draw red line at}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis, automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0 to 1, useful for dense plots}

\item{title}{optional string for plot title}

\item{color1}{first alternating color}

\item{color2}{second alternating color}

\item{annotate_var}{list of variables to annotate}

\item{annotate_p}{pvalue threshold to annotate}

\item{highlight_var}{list of variables to highlight}

\item{highlight_p}{pvalue threshold to highlight}

\item{highlighter}{color to highlight}

\item{groupcolors}{named list of colors for data in 'Color' column}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{res}{resolution of plot in pixels per inch}
}
\value{
png image(s)

png image(s)
}
\description{
Create Manhattan plots for EWAS
Note: There is an issue with dev.off() if using RStudio
Dependencies: ggplot2
Suggested: RColorBrewer

Create Manhattan plots for EWAS
Note: There is an issue with dev.off() if using RStudio
Dependencies: ggplot2
Suggested: RColorBrewer
}
\examples{
eman(d, format, groups, line, title=NULL, morecolors=FALSE, file="eman", hgt=7, wi=12, res=300 )
eman(d, format, groups, line, title=NULL, morecolors=FALSE, file="eman", hgt=7, wi=12, res=300 )
}
