% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ieman.R
\name{ieman}
\alias{ieman}
\title{ieman}
\usage{
ieman(d, line, log10 = TRUE, yaxis, ymax, opacity = 1, title = NULL,
  highlight_var, highlight_p, highlighter = "red", color1 = "#AAAAAA",
  color2 = "#4D4D4D", groupcolors, db, moreinfo = FALSE,
  background = "variegated", grpblocks = FALSE, file = "ieman", hgt = 7,
  wi = 12, bigrender = FALSE)
}
\arguments{
\item{d}{data frame, columns are Variable, pvalue, and Group; Shape, Color, and Info optional}

\item{line}{optional pvalue threshold to draw red line at}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis, automatically set if log10=TRUE}

\item{ymax}{set the upper limit for the y-axis if not automatically scaled}

\item{opacity}{opacity of points, from 0 to 1, useful for dense plots}

\item{title}{optional string for plot title}

\item{highlight_var}{vector of variables to highlight}

\item{highlight_p}{pvalue threshold to highlight}

\item{highlighter}{color to highlight}

\item{color1}{first alternating color}

\item{color2}{second alternating color}

\item{groupcolors}{named vector of colors for data in 'Color' column}

\item{db}{query address, ex. "https://www.google.com/search?q="}

\item{moreinfo}{includes more information on hover, refers to Info column}

\item{background}{variegated or white}

\item{grpblocks}{boolean, turns on x-axis group marker blocks}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{bigrender}{can set to TRUE for big plots (~50000 rows) that produce huge input lookup error}
}
\value{
png image(s)
}
\description{
Create interactive Manhattan plots for EWAS
}
\examples{
data(ewas)
ewas$Info <- paste0("Shape: ", ewas$Shape, "\\np-value: ", signif(ewas$pvalue, digits=3))
ieman(d=ewas, title="EWAS Example", line=0.001, color1="#A23B72", color2="#2A84AA",
moreinfo=TRUE, highlight_p=0.001, db="https://www.google.com/search?q=", highlighter="green")
}
\seealso{
\code{\link{eman}}, \code{\link{aeman}}, \code{\link{igman}}, \code{\link{ipheman}}

Other EWAS functions: \code{\link{aeman}},
  \code{\link{eman}}
}
