% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igman.R
\name{igman}
\alias{igman}
\title{igman}
\usage{
igman(d, line, log10 = TRUE, yaxis, opacity = 1, highlight_snp, highlight_p,
  highlighter = "red", title = NULL, chrcolor1 = "#AAAAAA",
  chrcolor2 = "#4D4D4D", groupcolors, db, background = "variegated",
  chrblocks = FALSE, moreinfo = FALSE, file = "igman", hgt = 7,
  wi = 12, bigrender = FALSE)
}
\arguments{
\item{d}{data frame, must contain SNP, CHR, POS, pvalue columns, optional Shape, Color, and Info}

\item{line}{optional pvalue threshold to draw red line at}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis, automatically set if log10=TRUE}

\item{opacity}{opacity of points, from 0 to 1, useful for dense plots}

\item{highlight_snp}{vector of SNPs to highlight}

\item{highlight_p}{pvalue threshold to highlight}

\item{highlighter}{color to highlight}

\item{title}{optional string for plot title}

\item{chrcolor1}{first alternating color for chromosome}

\item{chrcolor2}{second alternating color for chromosome}

\item{groupcolors}{named vector of colors for data in 'Color' column}

\item{db}{choose database to connect to ("dbSNP", "GWASCatalog", or enter your own search address)}

\item{background}{variegated or white}

\item{chrblocks}{boolean, turns on x-axis chromosome marker blocks}

\item{moreinfo}{includes more information on hover, refers to Info column}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{bigrender}{can set to TRUE for big plots (~50000 rows) that produce huge input lookup error}
}
\value{
html file
}
\description{
Create Interactive Manhattan plots for GWAS
}
\examples{
data(gwas)
#For this example we will change the "Frame" column to
#"Info" and set the flag moreinfo=TRUE to plot the information
#about the encoding method
colnames(gwas)[5] <- "Info"
igman(d=gwas, line=0.0005, title="GWAS Example", moreinfo=TRUE, db="dbSNP")
}
\seealso{
\code{\link{gman}}, \code{\link{agman}}, \code{\link{ipheman}}, \code{\link{ieman}}

Other GWAS functions: \code{\link{agman}},
  \code{\link{gman}}

Other interactive plotting functions: \code{\link{ipheman}}
}
