% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metro_lines}
\alias{metro_lines}
\title{All WMATA Rail Lines}
\format{
A tibble with 6 rows and 6 variables:
\describe{
\item{LineCode}{Two-letter abbreviation for the line (e.g.: RD, BL, YL, OR,
GR, or SV).}
\item{DisplayName}{Full name of line color.}
\item{StartStationCode}{Start station code. For example, will be F11
(Branch Avenue) for the Green Line, A15 (Shady Grove) for the Red Line,
etc. Use this value in other rail-related APIs to retrieve data about a
station.}
\item{EndStationCode}{End station code. For example, will be E10
(Greenbelt) for the Green Line, B11 (Glenmont) for the Red Line, etc. Use
this value in other rail-related APIs to retrieve data about a station.}
\item{InternalDestination}{Intermediate terminal station code(s). During
normal service, some trains on some lines might end their trip prior to the
\code{StartStationCode} or \code{EndStationCode.} A good example is on the Red Line
where some trains stop at A11 (Grosvenor) or B08 (Silver Spring). \code{NA} if
not defined.}
}
}
\usage{
metro_lines
}
\description{
All WMATA Rail Lines
}
\details{
As of 2021-02-22 22:00:00 ET.
}
\keyword{datasets}
