% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metro_routes}
\alias{metro_routes}
\title{All WMATA bus routes}
\format{
A data frame with 325 rows and 3 variables:
\describe{
\item{route}{Unique identifier for a given route variant.}
\item{name}{Descriptive name of the route variant.}
\item{description}{Denotes the route variant’s grouping – lines are a
combination of routes which lie in the same corridor and which have
significant portions of their paths along the same roadways.}
...
}
}
\usage{
metro_routes
}
\description{
All WMATA bus routes
}
\keyword{datasets}
