% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metro_stations}
\alias{metro_stations}
\title{All WMATA Rail Stations}
\format{
A tibble with 95 rows and 10 variables:
\describe{
\item{StationCode}{Station code for this station. Use this value in other
rail-related APIs to retrieve data about a station.}
\item{StationName}{Full name for this station, as shown on the WMATA
website.}
\item{StationTogether}{For stations with multiple platforms (e.g.: Gallery
Place, Fort Totten, L'Enfant Plaza, and Metro Center), the additional
\code{StationCode} will be listed here.}
\item{LineCodes}{Character vector of two-letter abbreviations (e.g.: RD,
BL, YL, OR, GR, or SV) served by this station. If the station has an
additional platform, the lines served by the other platform are listed in
the \code{LineCodes} values for the record associated with the \code{StationCode}
found in \code{StationTogether.}}
\item{Lat}{Latitude.}
\item{Lon}{Longitude.}
\item{Street}{Street address (for GPS use).}
\item{City}{City.}
\item{State}{State (abbreviated).}
\item{Zip}{Zip code.}
}
}
\usage{
metro_stations
}
\description{
All WMATA Rail Stations
}
\details{
As of 2021-02-22 22:00:00 ET.
}
\keyword{datasets}
