% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metro_stops}
\alias{metro_stops}
\title{All WMATA bus stops}
\format{
A data frame with 9,074 rows and 5 variables:
\describe{
\item{stop}{Bus stop ID}
\item{name}{Bus stop name}
\item{lon}{Latitude.}
\item{lat}{Longitude.}
\item{routes}{Character vector of routes services by stop.}
...
}
}
\usage{
metro_stops
}
\description{
All WMATA bus stops
}
\keyword{datasets}
