% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next-bus.R
\name{next_bus}
\alias{next_bus}
\title{Next buses}
\format{
A tibble with 6 variables:
\describe{
\item{route}{Base route name as shown on the bus.}
\item{direction}{Friendly description of direction and destination.}
\item{dir_num}{Denotes a binary direction (0 or 1) of the bus. There is no
specific mapping to direction, but a different value for the same route
signifies that the buses are traveling in opposite directions.}
\item{minutes}{Minutes until bus arrival at this stop.}
\item{vehicle}{Bus identifier.}
\item{trip}{Trip identifier.}
}
}
\usage{
next_bus(stop = NULL)
}
\arguments{
\item{stop}{7-digit regional stop ID.}
}
\description{
Returns next bus arrival times at a stop.
}
