% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheman.R
\name{pheman}
\alias{pheman}
\title{pheman}
\usage{
pheman(d, format = "plotman", groups = NULL, line, annotate_snp, annotate_p,
  title = NULL, morecolors = FALSE, file = "gman", hgt = 7, wi = 12,
  res = 300)
}
\arguments{
\item{d}{data frame, if not plato or plink format, must contain PHE, SNP, CHR, POS, pvalue columns, optional Shape}

\item{format}{format of input}

\item{groups}{optional file for organizing phenotypes}

\item{line}{optional pvalue threshold to draw red line at}

\item{annotate_snp}{list of RSIDs to annotate}

\item{annotate_p}{pvalue threshold to annotate}

\item{title}{optional string for plot title}

\item{morecolors}{boolean, expand color palette, requires RColorBrewer package, default FALSE}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{res}{resolution of plot in pixels per inch}
}
\value{
png image
}
\description{
Create Manhattan plots for GWAS
Note: There is an issue with dev.off() if using RStudio
Dependencies: ggplot2
Suggested: RColorBrewer, ggrepel
}
\examples{
pheman(d, format, line, groups, annotate_snp, annotate, title, morecolors, file, hgt, wi, res)
}
