% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheman.R
\name{pheman}
\alias{pheman}
\title{pheman}
\usage{
pheman(d, phegroup, line, log10 = TRUE, yaxis, ymax, opacity = 1,
  annotate_snp, annotate_p, highlight_snp, highlight_p, highlighter = "red",
  title = NULL, chrcolor1 = "#AAAAAA", chrcolor2 = "#4D4D4D", groupcolors,
  background = "variegated", chrblocks = TRUE, file = "pheman", hgt = 7,
  wi = 12, res = 300)
}
\arguments{
\item{d}{data frame, must contain PHE, SNP, CHR, POS, pvalue columns, optional Shape}

\item{phegroup}{optional grouping file for phenotypes, must contain PHE and Group columns}

\item{line}{optional pvalue threshold to draw red line at}

\item{log10}{plot -log10() of pvalue column, boolean}

\item{yaxis}{label for y-axis, automatically set if log10=TRUE}

\item{ymax}{set the upper limit for the y-axis if not automatically scaled}

\item{opacity}{opacity of points, from 0-1, useful for dense plots}

\item{annotate_snp}{vector of SNPs to annotate}

\item{annotate_p}{pvalue threshold to annotate}

\item{highlight_snp}{vector of snps to highlight}

\item{highlight_p}{pvalue threshold to highlight}

\item{highlighter}{color to highlight}

\item{title}{optional string for plot title}

\item{chrcolor1}{first alternating color for chromosome}

\item{chrcolor2}{second alternating color for chromosome}

\item{groupcolors}{named vector of colors for data in 'PHE' or 'Group' column}

\item{background}{variegated or white}

\item{chrblocks}{boolean, turns on x-axis chromosome marker blocks}

\item{file}{file name of saved image}

\item{hgt}{height of plot in inches}

\item{wi}{width of plot in inches}

\item{res}{resolution of plot in pixels per inch}
}
\value{
png image
}
\description{
Create Manhattan plots for PheWAS
}
\examples{
data(phewas)
pheman(d=phewas, title="PheWAS Example")
}
\seealso{
\code{\link{ipheman}}, \code{\link{apheman}}, \code{\link{gman}}, \code{\link{eman}}

Other PheWAS functions: \code{\link{apheman}},
  \code{\link{ipheman}}

Other static plotting functions: \code{\link{eman}},
  \code{\link{gman}}
}
