% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-dest.R
\name{rail_destination}
\alias{rail_destination}
\title{Station to station}
\usage{
rail_destination(from = NULL, to = NULL)
}
\arguments{
\item{from}{Station code for the origin station.}

\item{to}{Station code for the destination station.}
}
\description{
Returns a distance, fare information, and estimated travel time between any
two stations, including those on different lines. Omit both parameters to
retrieve data for all stations. Omit one parameter to return all combinations
to or from the provided station.
}
\details{
\itemize{
\item \code{peak_fee}: Fare during peak times (weekdays from opening to 9:30 AM and
3-7 PM, and weekends from midnight to closing).
\item \code{off_fee}: Fare during off-peak times.
\item \code{min_fee}: Reduced fare for senior citizens or people with disabilities.
}
}
