% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rail-incidents.R
\name{rail_incidents}
\alias{rail_incidents}
\title{Rail incidents}
\usage{
rail_incidents()
}
\description{
Returns reported rail incidents (significant disruptions and delays to normal
service). The data is identical to WMATA's Metrorail Service Status feed.
}
