% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-stations.R
\name{rail_stations}
\alias{rail_stations}
\title{Metro rail stations}
\usage{
rail_stations(line = NULL)
}
\arguments{
\item{line}{Two-letter line code abbreviation, see Details or \code{\link[=rail_lines]{rail_lines()}}.
If \code{NULL}, (the default) all stations are returned.}
}
\description{
The stations on a Metro rail line, including the full names and geographic
locations. The four big transfer stations have two station code designations,
one for each track direction. All the lines running through each station
are listed in a single list-column.
}
\details{
The \code{line_code} takes one of these two-letter codes:
\itemize{
\item BL = Blue
\item GR = Green
\item OR = Orange
\item RD = Red
\item SV = Silver
\item YL = Yellow
}

The result of this endpoint is saved as the \link{stations} object.
}
