% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rail-times.R
\name{rail_times}
\alias{rail_times}
\title{Rail Station Timings}
\format{
A tibble 1 row per train with 8 variables:
\describe{
\item{StationCode}{Station code for this station. Use this value in other
rail-related APIs to retrieve data about a station.}
\item{StationName}{Full name of the station.}
\item{Weekday}{Abbreviated day of the week.}
\item{OpeningTime}{Station opening time (EST). Format is HH:mm.}
\item{FirstStation}{Station code for the train's destination. Use this
value in other rail-related APIs to retrieve data about a station.}
\item{FirstTime}{First train leaves the station at this time (EST). Format
is HH:mm.}
\item{LastStation}{Station code for the train's destination. Use this value
in other rail-related APIs to retrieve data about a station.}
\item{LastTime}{Last train leaves the station at this time (EST). Format is
HH:mm. Note that when the time is AM, it signifies the next day. For
example, a value of 02:30 under a Saturday element means the last train
leaves on Sunday at 2:30 AM.}
}
}
\usage{
rail_times(StationCode = NULL)
}
\arguments{
\item{StationCode}{Station code. Use the \code{\link[=rail_stations]{rail_stations()}} function to return
a list of all station codes. Use \code{NULL} (default) to return times for all
stations.}
}
\value{
A \emph{tidy} data frame of station schedules. Combined from a nested list
of weekday times.
}
\description{
Returns opening and scheduled first/last train times based on a given
\code{StationCode}. Omit the \code{StationCode} to return timing information for all
stations.
}
\details{
Note that for stations with multiple platforms (e.g.: Metro Center, L'Enfant
Plaza, etc.), a distinct call is required for each \code{StationCode} to retrieve
the full set of train times at such stations.
}
\examples{
\dontrun{
rail_times("A01")
}
}
\seealso{
\url{https://developer.wmata.com/docs/services/5476364f031f590f38092507/operations/5476364f031f5909e4fe3312}

Other Rail Station Information: 
\code{\link{rail_destination}()},
\code{\link{rail_entrance}()},
\code{\link{rail_lines}()},
\code{\link{rail_path}()},
\code{\link{rail_stations}()}
}
\concept{Rail Station Information}
