% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wmata_api}
\alias{wmata_api}
\title{Call the WMATA API}
\usage{
wmata_api(
  type = c("Rail", "Bus", "Incidents", "NextBusService", "StationPrediction"),
  endpoint,
  query = NULL
)
}
\arguments{
\item{type}{The API base type to call, one of "Rail" or "Bus".}

\item{endpoint}{The API endpoint (e.g., "jStations").}

\item{query}{Additional queries also passed, possibly your key if need be.}
}
\value{
A single JSON string.
}
\description{
Use \code{\link[httr:GET]{httr::GET()}} to make a request to the WMATA API and return the unparsed
JSON string. This function is used in others devoted to a
specific APIs and endpoints.
}
\details{
This function was modified from 'zamorarr/wmata' on GitHub:
\url{https://github.com/zamorarr/wmata/blob/master/R/api.r}
}
\examples{
\dontrun{
wmata_api("Rail", "jLines", query = list(LineCode = "RD"))
}
}
\keyword{internal}
