% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{wmata_key}
\alias{wmata_key}
\alias{wmata_demo}
\alias{browse_wmata_key}
\title{Find a WMATA API key}
\usage{
wmata_key()

wmata_demo()

browse_wmata_key()
}
\description{
All calls to the WMATA API must be accompanied by a personal API key. A free
key can be obtained by subscribing to the default tier. See
\code{\link[=browse_wmata_key]{browse_wmata_key()}} for more information on getting a key and storing on
your system.

The WMATA also provides a demonstration key. This key should \strong{never} be
used in production applications, it is rate limited and subject to change at
\emph{any} time. The key can be found on
\href{https://developer.wmata.com/products/5475f236031f590f380924ff}{this page}.
If the user has the "rvest" package installed, the \code{\link[=wmata_demo]{wmata_demo()}} function
can be used to try scrape said page and look for the key automatically.
}
\details{
Default tier sufficient for most casual developers. Rate limited to 10
calls/second and 50,000 calls per day. This product contains 8 APIs:
\itemize{
\item Bus Route and Stop Methods
\item GTFS
\item Incidents
\item Misc Methods
\item Rail Station Information
\item Real-Time Bus Predictions
\item Real-Time Rail Predictions
\item Train Positions
}
}
